backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
langgraph
google-generativeai
python-dotenv
sse-starlette
httpx
pydantic>=2.0.0
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
aiohttp
